requirements.txt
fastapi
uvicorn
weaviate-client
sentence-transformers
pydantic